[#922] Refactoring react folder structure
with @anikarni
