+++ web-ui/src/account_recovery/account_recovery.js
@@ -3,8 +3,8 @@ import { render } from 'react-dom'
 import a11y from 'react-a11y'
 import { I18nextProvider } from 'react-i18next'
 
-import Page from 'js/account_recovery/page'
-import i18n from 'js/account_recovery/i18n'
+import Page from './page'
+import i18n from '../i18n'
 
 import 'font-awesome/scss/font-awesome.scss'
