--- web-ui/test/jasmine.json
@@ -1,9 +1,0 @@
-{
-  "spec_dir": "test/spec/account_recovery/",
-  "spec_files": [
-    "**/*[sS]pec.js"
-  ],
-  "helpers": [
-    "helpers/**/*.js"
-  ]
-}
