+++ web-ui/webpack.config.js
@@ -7,7 +7,7 @@ var aliases = require('./config/alias-webpack');
 module.exports = {
   entry: {
     app: './app/js/index.js',
-    account_recovery: './app/js/account_recovery.js',
+    account_recovery: './src/account_recovery/account_recovery.js',
     sandbox: './app/js/sandbox.js'
   },
   node: { fs: 'empty' },
